rename a few classes
