+++ src/Pim/Attribute/Attribute.php
@@ -7,51 +7,29 @@ declare(strict_types=1);
 
 namespace Wizaplace\SDK\Pim\Attribute;
 
-abstract class Attribute
+final class Attribute extends AbstractAttribute
 {
-    /** @var int */
-    private $id;
-
-    /** @var string */
-    private $name;
-
-    /** @var AttributeType */
-    private $type;
-
-    /** @var int[] */
-    private $categoriesPath;
+    /** @var AttributeVariant[]|null */
+    private $variants;
 
     /**
      * @internal
      */
     public function __construct(array $data)
     {
-        $this->id = $data['feature_id'];
-        $this->name = $data['description'];
-        $this->type = new AttributeType($data['feature_type']);
-        $this->categoriesPath = $data['categories_path'];
-    }
-
-    public function getId(): int
-    {
-        return $this->id;
-    }
-
-    public function getName(): string
-    {
-        return $this->name;
-    }
-
-    public function getType(): AttributeType
-    {
-        return $this->type;
+        parent::__construct($data);
+        if (isset($data['variants'])) {
+            $this->variants = array_map(function (array $variantData): AttributeVariant {
+                return new AttributeVariant($variantData);
+            }, $data['variants']);
+        }
     }
 
     /**
-     * @return int[]
+     * @return AttributeVariant[]
      */
-    public function getCategoriesPath(): array
+    public function getVariants(): ?array
     {
-        return $this->categoriesPath;
+        return $this->variants;
     }
 }
